+++ core/process_metadata.py
@@ -4,7 +4,7 @@
 # ------------------------------------------------------------------
 
 
-from api.bangumi_model import SubjectRelation
+from api.bangumi_model import BangumiBaseType, SubjectPlatform, SubjectRelation
 from api.komga_api import *
 from pypinyin import slug, Style
 
@@ -32,7 +32,7 @@ def _set_genres(komga_metadata, bangumi_metadata):
     if bangumi_metadata["platform"] is None:
         genrelist.append("其他")
     else:
-        genrelist.append(bangumi_metadata["platform"])
+        genrelist.append(SubjectPlatform.parse(bangumi_metadata["platform"]).cn)
     for info in bangumi_metadata["infobox"]:
         if info["key"] == "连载杂志":
             if type(info["value"]) == list:
@@ -54,6 +54,7 @@ def _set_status(komga_metadata, bangumi_metadata):
     runningLang = ["放送", "放送（連載）中"]
     abandonedLang = ["打ち切り"]
     endedLang = ["完結", "结束", "连载结束"]
+    hiatusLang = ["有生之年"]
 
     casestatus = "ONGOING"
 
@@ -66,6 +67,9 @@ def _set_status(komga_metadata, bangumi_metadata):
         elif info["key"] in endedLang:
             casestatus = "ENDED"
             break
+        elif info["key"] in hiatusLang:
+            casestatus = "HIATUS"
+            break
 
     komga_metadata.status = casestatus
 
@@ -73,6 +77,7 @@ def _set_status(komga_metadata, bangumi_metadata):
 def _set_total_book_count(komga_metadata, subject_relations):
     """
     漫画总册数
+    bgm 元数据中的 volumes 字段不一定准确/更新及时，改为根据关联条目统计
     """
     totalBookCount = 0
     for relation in subject_relations:
@@ -104,6 +109,7 @@ def _set_language(komga_metadata, manga_filename):
                 "天下",
                 "青文",
                 "长鸿",
+                "長鴻",
                 "角川",
                 "文传",
                 "文傳",
@@ -114,6 +120,7 @@ def _set_language(komga_metadata, manga_filename):
     for langCode, keywords in languageTypes:
         if any(keyword in manga_filename for keyword in keywords):
             komga_metadata.language = langCode
+
 
 def _set_alternate_titles(komga_metadata, bangumi_metadata):
     """
@@ -226,19 +233,25 @@ def _set_links(komga_metadata, bangumi_metadata, subject_relations):
             "url": "https://bgm.tv/subject/" + str(bangumi_metadata["id"]),
         }
     ]
-    for relation in subject_relations:
-        if relation["relation"] == "动画":
-            link = {
-                "label": "动画：" + relation["name"],
-                "url": "https://bgm.tv/subject/" + str(relation["id"]),
-            }
-            links.append(link)
-        if relation["relation"] == "书籍":
-            link = {
-                "label": "书籍：" + relation["name"],
-                "url": "https://bgm.tv/subject/" + str(relation["id"]),
-            }
-            links.append(link)
+    if subject_relations:
+        # 添加改编动画、书籍链接
+        for relation in subject_relations:
+            if (
+                SubjectRelation.parse(relation["relation"]).value
+                == SubjectRelation.ADAPTATION.value
+            ):
+                if BangumiBaseType.parse(relation["type"]) == BangumiBaseType.ANIME:
+                    link = {
+                        "label": "动画：" + relation["name"],
+                        "url": "https://bgm.tv/subject/" + str(relation["id"]),
+                    }
+                    links.append(link)
+                elif BangumiBaseType.parse(relation["type"]) == BangumiBaseType.BOOK:
+                    link = {
+                        "label": "书籍：" + relation["name"],
+                        "url": "https://bgm.tv/subject/" + str(relation["id"]),
+                    }
+                    links.append(link)
     komga_metadata.links = links